// ui/src/lib/api-client/models/GetFhirServersByServerName200Response.ts
export interface GetFhirServersByServerName200Response {
    /**
     * Server identifier used in URLs
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    name: string;
    /**
     * Human-readable server name from FHIR metadata
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    displayName: string;
    /**
     * Original server URL
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    url: string;
    /**
     * FHIR version supported by server
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    fhirVersion: string;
    /**
     * Server software version
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    serverVersion?: string;
    /**
     * Server software name from FHIR metadata
     * @type {string}
     * @memberof GetFhirServersByServerName200Response
     */
    serverName?: string;
    /**
     * Whether this server is supported
     * @type {boolean}
     * @memberof GetFhirServersByServerName200Response
     */
    supported: boolean;
    /**
     * 
     * @type {GetFhirServers200ResponseServersInnerEndpoints}
     * @memberof GetFhirServersByServerName200Response
     */
    endpoints: GetFhirServers200ResponseServersInnerEndpoints;
}